Runic::Parser setting to (dis)allow top level expressions
Top level expressions are invalid when a file is compiled to object
code, so they must be disallowed. Yet, in some cases we may want
top level expressions to be parsed, for example in the test suite
or a REPL.

This patch introduces a `top_level_expressions: bool` setting to
Runic::Parser to allow or disable them (default).