Initial interactive session
Running `bin/runic interactive` starts an interactive session,
during which externs, defs and a single expression will be
evaluated as they are typed.

You may pass `--debug` to print the LLVM IR as it is generated.

The session is limited to a single expression (for the time being).
You may not assign variables then later access them for example.