Default arguments
Functions and methods can now have a default value. The value must
be a literal, and the argument type will be inferred to be the
literal type unless specified.

Futhermore literals are now casted when possible in calls and
definitions. For example a method expecting a f64 but receiving an
i32 literal will automatically cast the i32 literal into a f64.

Functions and methods still expect all arguments to be fulfilled,
and semantic analysis is responsible for injecting the default
values directly into the function/method calls. This avoids
nilables, overloads or another complexity.

Examples:

```runic
def foo(a : int, b = 0, c : f32 = 1.2)
end

foo(1) # => foo(1, 0, 1.2_f32)
foo(1, 2, 1) # => foo(1, 2, 1.0_f32)
```