Fix: lexer distinguishes linefeeds from semicolons
Sometimes the parser must be capable to allow linefeeds but disallow
the presence of semicolons. For example in argument definitions or
within array and hash literals.

The lexer thus have to return different tokens for both cases, yet
still group them when needed, since a semicolon is still considered
as a line delimiter. When encountering a semicolon we keep grouping
all whitespace, linefeeds and semicolons that immediately follow.
When encountering a linefeed, we only group the whitespace and
linefeeds that follow, stopping when we find a semicolon.