Semantic: resolve operators from struct methods
Replaces the builtin hardcoded definitions for math, logical and
bitwise operators to be defined as struct methods instead, written
directly in Runic.

This will allow custom structs to implement these operators, and
already allows to document operators as simple methods, and to have
different behaviors by switching to another corelib or by reopening
structs to support operators on different types, for example.