Fix: lex and parse -number.method and integer.method
Modifies the lexer to stop expecting a float when an integer is
followed by a dot, since it may be a method call, for example `1.abs`.

Modifies the parser so the unary sign is set on the following
number, instead of the following expression, so `-1.abs` is parsed
`(-1).abs` not `-(1.abs)`. Note that `-x.abs` is still parsed as
`-(x.abs)` since calls have a higher precedence than unary
expressions.