Semantic analysis of struct initialize method
Verifies that the initialize method of structs do initialize all the
instance variables of the struct.

Verifies that ivars are initialized in all control-flow branches,
and allows accessing an ivar inside the initialize method as long as
it has been initialized in the current scope/branch.

Outside of initialize methods, instance variables are assumed to
always have been initialized, thus have a valid value.