Removed abstract method return types
The crystal compiler now complains when implementing abstract
methods, forcing to specify the return type for each and every
implementation whenever the type is specified

Since typing every single implementation is boring and doesn't bring
any benefit, this patch removes the return type of abstract methods.

Why not merely check that implementations are indeed returning the
expected type, or discarding it when the abstract method definition
is typed as Nil is a mystery.